requirements.txt
jinja2
PyYAML
requests
psycopg2